+++ app/src/components/fields/expandedTextSection/expandedTextSection.tsx
@@ -17,20 +17,20 @@
 import { useState, useRef, useEffect } from 'react';
 import classNames from 'classnames/bind';
 import { useIntl } from 'react-intl';
-import { messages } from './messages';
+import { messages } from 'pages/inside/productVersionPage/linkedTestCasesTab/tagList/messages';
 import styles from './expandedTextSection.scss';
 
 const cx = classNames.bind(styles);
 
 interface ExpandedTextSectionProps {
   text: string;
-  defaultVisibleLines: number;
+  defaultVisibleLines?: number;
   fontSize?: number;
 }
 
 export const ExpandedTextSection = ({
   text,
-  defaultVisibleLines,
+  defaultVisibleLines = 3,
   fontSize = 13,
 }: ExpandedTextSectionProps) => {
   const { formatMessage } = useIntl();
